// Front-End/src/app/auth/interfaces/atractivo.interface.ts
export interface Atractivo {
  id: string;
	nombre: string;
	categoria: string;
	tipologia: string;
	localizacion: string;
	descripcion: string;
	accesibilidad: string;
	actividades: string;
	num_visitante_ideal: string;
	fecha_ideal_visita: string;
	segmento_mercado_potencial: string;
	costo: string;
	servicio: string;
  alt_img?: string;
  selected?: boolean;
}